Support multiple fields on the same page
We weren't being careful enough with our selectors previously to all for
multiple fields on the same page. They were getting confused,
particularly with the options and then the selector code in the
image_cropping section. We make sure we pass the field name around and
use it where necessary.