Allow use of custom CSS
I've added this so that I don't have to make changes to the actual CSS
here get the look that I want. I'm not sure what the best way forward
is. In the end I copy & pasted all the current css and then tweaked it
in my own project and enabled this option.