Use 'data-backdrop="static"' for modal
So that clicking on the background doesn't exit the modal. The issue
here is that it is easy to end up clicking on the background when
dragging the crop window around and it ends up dismissing the modal in
an annoying way. This solves the issue by changing it so that you can
only dismiss the modal by using the 'close' button.