Remove try/excpet & raise on same width
I encountered an issue when trying to upload an image which was already
the correct size. For some reason we were just raising nothing if we
encountered that. That doesn't seem to be necessary so I've removed
that.

I've also removed the whole try catch as I don't believe it was doing
much good and was just masking the real error when in development.