Add option to clear file upload when editing
If there is already a filename value then we show a 'clear' button which
checks a hidden checkbox which follows the naming convention of the
ClearableFieldField from Django.

When we add a new file then we uncheck the check box. We only show the
button when there is a field value.